mobile/requirements.txt
Appium-Python-Client>=3.1.0
openai>=1.0.0
groq>=0.9.0
